requirements.txt
# Azure + ML
azureml-core
azureml-mlflow
azure-storage-blob
azure-identity
azure-ai-ml



# TensorFlow stack (ensure compatibility with MLflow)
tensorflow==2.13.0
keras==2.13.1


# MLflow (must be compatible with Azure ML + TF 2.13)
mlflow==2.10.1

# Core dependencies
numpy==1.24.3
typing-extensions==4.5.0

# Image Processing & Augmentations
pydicom
opencv-python
albumentations==1.3.1
google-api-core==2.25.1

# API
Flask
requests
gunicorn
python-dotenv
